Moved CachedSourceAvailability to not be an inner class
