+++ catalog/core/catalog-core-api-impl/src/main/java/ddf/catalog/util/impl/SourcePoller.java
@@ -21,7 +21,6 @@ import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.RemovalNotification;
 import com.google.common.primitives.Ints;
-import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.MoreExecutors;
 import ddf.catalog.source.CatalogProvider;
 import ddf.catalog.source.CatalogStore;
@@ -31,26 +30,19 @@ import ddf.catalog.source.Source;
 import ddf.catalog.source.SourceMonitor;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.Callable;
-import java.util.concurrent.CancellationException;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang3.time.DateUtils;
 import org.codice.ddf.platform.util.StandardThreadFactoryBuilder;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
@@ -226,7 +218,8 @@ public class SourcePoller implements EventListenerHook {
           final CachedSourceAvailability cachedSourceAvailability =
               cachedSourceStatusOptional.get();
           cachedSourceAvailability.cancel();
-          cachedSourceAvailability.recheckAvailability(source, pollIntervalMinutes);
+          cachedSourceAvailability.recheckAvailability(
+              source, pollIntervalMinutes, sourceAvailabilityChecksThreadPool);
         } else {
           LOGGER.debug(
               "Even though the source id={} has already been bound, the SourceKey for the source has been changed. The entry for the old SourceKey will be invalidated and a new entry will be bound at the next handleAllSources call.",
@@ -361,12 +354,14 @@ public class SourcePoller implements EventListenerHook {
         LOGGER.trace(
             "Starting a process to recheck the availability of source id={}, which has already been added to the cache ",
             source.getId());
-        cachedSourceAvailability.recheckAvailability(source, pollIntervalMinutes);
+        cachedSourceAvailability.recheckAvailability(
+            source, pollIntervalMinutes, sourceAvailabilityChecksThreadPool);
       } else {
         LOGGER.debug("Found a new SourceKey for source id={}", source.getId());
         final CachedSourceAvailability cachedSourceAvailability = new CachedSourceAvailability();
         cache.put(sourceKey, cachedSourceAvailability);
-        cachedSourceAvailability.recheckAvailability(source, pollIntervalMinutes);
+        cachedSourceAvailability.recheckAvailability(
+            source, pollIntervalMinutes, sourceAvailabilityChecksThreadPool);
       }
     }
 
@@ -380,136 +375,6 @@ public class SourcePoller implements EventListenerHook {
         .orElseThrow(
             () ->
                 new IllegalStateException("Unable to get the bundle context for the SourcePoller"));
-  }
-
-  /**
-   * Used to keep track of the last known {@link SourceAvailability} in the {@link #cache} and to
-   * cancel the current availability check
-   */
-  private class CachedSourceAvailability {
-
-    /** {@link Optional#empty()} indicates that an availability check has not yet completed */
-    private Optional<SourceAvailability> sourceAvailabilityOptional;
-
-    private Future<Boolean> isAvailableFuture;
-
-    private Future watcherFuture;
-
-    public CachedSourceAvailability() {
-      sourceAvailabilityOptional = Optional.empty();
-      isAvailableFuture = Futures.immediateCancelledFuture();
-      watcherFuture = Futures.immediateCancelledFuture();
-    }
-
-    private void cancel() {
-      watcherFuture.cancel(true);
-      isAvailableFuture.cancel(true);
-      sourceAvailabilityOptional = Optional.empty();
-    }
-
-    /**
-     * Starts a process to recheck the availability if it is not currently being checked and if has
-     * not been checked within the last {@code timeoutMinutes} minutes.
-     */
-    public void recheckAvailability(final Source source, final int timeoutMinutes) {
-      notNull(source);
-      final String sourceId = source.getId();
-
-      if (!watcherFuture.isDone()) {
-        LOGGER.debug(
-            "A process has already been started to check the availability of source id={}. This is expected when the source is recently modified or bound. Not starting a new check",
-            sourceId);
-        return;
-      }
-
-      if (!isAvailableFuture.isDone()) {
-        isAvailableFuture.cancel(true);
-      }
-
-      if (sourceAvailabilityOptional.isPresent()) {
-        final Date sourceStatusDate = sourceAvailabilityOptional.get().getSourceStatusDate();
-        final Date nextEarliestRecheckDate = DateUtils.addMinutes(sourceStatusDate, timeoutMinutes);
-        final Date now = new Date();
-
-        if (now.before(nextEarliestRecheckDate)) {
-          // Not rechecking the availability of source id={} because it has been checked in the last
-          // {} minutes
-          // Don't log anything here because this happens really often.
-          return;
-        }
-      }
-
-      LOGGER.trace("Starting a process to check the availability of source id={}", sourceId);
-      isAvailableFuture =
-          sourceAvailabilityChecksThreadPool.submit((Callable<Boolean>) source::isAvailable);
-      watcherFuture =
-          sourceAvailabilityChecksThreadPool.submit(
-              () -> watchAndUpdateAvailability(isAvailableFuture, sourceId, pollIntervalMinutes));
-    }
-
-    /**
-     * Availability checks will be cancelled if they take longer than the poll interval, and the
-     * {@link SourceStatus} will be updated to {@link SourceStatus#TIMEOUT}.
-     */
-    private void watchAndUpdateAvailability(
-        final Future<Boolean> isAvailableFuture, final String sourceId, long timeoutMinutes) {
-      SourceStatus newSourceStatus;
-
-      try {
-        final boolean isAvailable = isAvailableFuture.get(timeoutMinutes, TimeUnit.MINUTES);
-        LOGGER.trace("Successfully checked the availability of source id={}", sourceId);
-        newSourceStatus = isAvailable ? SourceStatus.AVAILABLE : SourceStatus.UNAVAILABLE;
-      } catch (TimeoutException e) {
-        isAvailableFuture.cancel(true);
-        LOGGER.debug(
-            "Unable to check the availability of source id={} within {} {}. Cancelling the check",
-            sourceId,
-            timeoutMinutes,
-            TimeUnit.MINUTES);
-        newSourceStatus = SourceStatus.TIMEOUT;
-      } catch (CancellationException e) {
-        isAvailableFuture.cancel(true);
-        LOGGER.debug("Unable to check the availability of source id={}", sourceId, e);
-        return;
-      } catch (RuntimeException | ExecutionException e) {
-        isAvailableFuture.cancel(true);
-        LOGGER.debug("Exception checking the availability of source id={}", sourceId, e);
-        newSourceStatus = SourceStatus.EXCEPTION;
-      } catch (InterruptedException e) {
-        isAvailableFuture.cancel(true);
-        LOGGER.debug("Interrupted while checking the availability of source id={}", sourceId);
-        Thread.currentThread().interrupt();
-        return;
-      }
-
-      if (sourceAvailabilityOptional.isPresent()) {
-        final SourceAvailability oldSourceAvailability = sourceAvailabilityOptional.get();
-
-        if (oldSourceAvailability.getSourceStatus() != newSourceStatus) {
-          LOGGER.info(
-              "The source status of source id={} has changed to {}. The last known source status was {} at {}.",
-              sourceId,
-              newSourceStatus,
-              oldSourceAvailability.getSourceStatus(),
-              oldSourceAvailability.getSourceStatusDate());
-        }
-
-        final SourceAvailability newSourceAvailability = new SourceAvailability(newSourceStatus);
-        sourceAvailabilityOptional = Optional.of(newSourceAvailability);
-      } else {
-        final SourceAvailability initialSourceAvailability =
-            new SourceAvailability(newSourceStatus);
-        LOGGER.info(
-            "Checked the availability of source id={} for the first time. Setting the source status to {}",
-            sourceId,
-            initialSourceAvailability.getSourceStatus());
-        sourceAvailabilityOptional = Optional.of(initialSourceAvailability);
-      }
-    }
-
-    public Optional<SourceAvailability> getSourceAvailability() {
-      return sourceAvailabilityOptional;
-    }
   }
 
   /**
